-Corrected mistake in Global.cs -Added GUI Code -Changed Global.MaxPlayers to a byte -Command spawns multitools for FB crews -Created GUI for setting player limits and role leads

+++ Max_Players/Mod.cs
@@ -15,7 +15,7 @@ namespace Max_Players
                     Global.rolelimits[i] = int.Parse(limits[i]);
                 }
             }
-            Global.MaxPlayers = PLXMLOptionsIO.Instance.CurrentOptions.GetStringValueAsInt("MaxPlayerLimit");
+            Global.MaxPlayers = (byte)PLXMLOptionsIO.Instance.CurrentOptions.GetStringValueAsInt("MaxPlayerLimit");
         }
 
         public override string Version => "0.8.2";
