+++ Max_Players/Players.cs
@@ -10,7 +10,7 @@ namespace Max_Players
 		// Token: 0x06000007 RID: 7 RVA: 0x000023A5 File Offset: 0x000005A5
 		private static void Prefix(ref RoomOptions roomOptions) //Load saved settings and set room limit to specified params
 		{
-			roomOptions.MaxPlayers = (byte)Global.MaxPlayers;
+			roomOptions.MaxPlayers = Global.MaxPlayers;
 		}
 	}
 }
